requirements.txt
# Core LLM multi-agent framework
autogen-agentchat[llm]~=0.2                       # multi-agent LLM communication framework :contentReference[oaicite:0]{index=0}
autogen-ext[openai]~=0.2                          # extension support for OpenAI API :contentReference[oaicite:1]{index=1}

# LLM clients
openai>=0.27.0                                     # OpenAI Python SDK (GPT-4, GPT-4o, etc)
python-dotenv>=1.0.0                               # Load environment variables (API keys etc)

# Tokenizers / embeddings
tiktoken>=0.4.0                                    # Tokenizer utility compatible with OpenAI
transformers>=4.30.0                               # Useful if you use local models or embeddings

# Data loading and benchmarks
datasets>=2.13.0                                   # HuggingFace datasets (FEVER, HotpotQA, etc.)

# Retrieval & IR (optional but useful)
faiss-cpu>=1.7.4                                   # Efficient similarity search (if using retrieval)
sentence-transformers>=2.2.2                       # Sentence embeddings for retrieval

# Scientific Python
numpy>=1.26.0
pandas>=2.1.0

# HTTP requests (for web retrieval / external sources)
requests>=2.30.0

# Optional evaluation utilities
scikit-learn>=1.3.0                                # metrics/precision/recall etc.
rouge-score>=0.1.2                                 # evaluation for QA/spans

# Logging / experiment tracking
rich>=13.0.0
tqdm>=4.65.0                                       # Progress bars for monitoring
tensorboard>=2.13.0                                # experiment tracking (optional)

# (optional) Tool execution if needed
psutil>=5.9.0                                      # system utilities
